[tricium spellchecker] Fix handling of block comment state
The current behavior of spellchecker was that if there is
no block comment pattern in the comment formats file, then
everything is considered to be in a block comment (since
block comment start is empty string if not specified).

This fixes it so that if block comment start is empty,
we should never enter block comment state.

Bug: 877486
Change-Id: Iebcd42d0231abe686d2580ad66e407b56c8a19ad
Reviewed-on: https://chromium-review.googlesource.com/1243681
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#17869}