[Findit] Remove invalid bug filing criteria in flake detector
This criteria was added to deal with the situation when flake detector
hits the bug filing limit and cases delays in reporting flakes.

It was valid because the old limit is 10 bugs per day, and flake
detector usually hits it, but now that the limit has been increased to
30, it's unlikely it will hit it anymore, so this criteria becomes
invalid.

Change-Id: I7fbf63c67cf5bdfa55f039b5628b5c4930bb5fdf
Reviewed-on: https://chromium-review.googlesource.com/1244283
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Reviewed-by: Jeffrey Li <lijeffrey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#17889}