[tricium] Re-add run URL with run ID in posted comments
In https://chromium-review.googlesource.com/1176443 I had removed
run ID and URL from comments posted, thinking that maybe the comments
would be posted but without the run ID being displayed, but it turned
out that run ID was required and comments weren't posted without it,
so run ID was re-added.

If we have to include run ID, maybe we'll include the URL as well,
since it might be useful. The expected effect is that the run ID
in comments will become links.

Bug: 852889
Change-Id: I0c1f139e76184a292be77fb928be2ebcfd769840
Reviewed-on: https://chromium-review.googlesource.com/1245732
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#17891}