[tricium pylint] Change enable/disable flag behavior
Background: The overall plan for configuration of checks in Pylint is
that eventually we probably want to generally use project-specific
pylintrc files for most projects; but it might be nice to support
an alternate simpler way to configure checks for a project, by
listing checks to enable in the config.

Most of the time I imagine that we would either want to just use
the default checks but disable a few specific checks, or make a
whitelist of checks to enable. This would make it so that the way
to do the former would be to just specify "disable", and the way
to do the latter would be to set disable to something general
like "all" or "W" and set enable to a specific list.

Bug: 864685
Change-Id: I57b0ac4439e13a5d0a0a809cbabe0fb4e97179d2
Reviewed-on: https://chromium-review.googlesource.com/1249955
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#17932}