[Findit] Modify cq false rejection query to handle retry with patch
Originally, the query relies on the (retry summary) step to tell if
test failures have caused the build to fail, however, after
(retry with patch) steps are added, the assumption is not valid anymore.

With (retry with patch) steps, test failures only fail a build if they
fail in (with patch) steps, pass (without patch) steps and fail again
in (retry with patch) steps, so, now the most reliable way to detect
test failures that have caused failure build is to check the
(retry with patch) steps.

Change-Id: Id12a7e14edded990da288e5a0143cd711d20aeb0
Reviewed-on: https://chromium-review.googlesource.com/1250404
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#17950}