[Findit] Save build_completed to WfAnalysis at the beginning of the analysis.
Before we save build_completed to WfAnalysis when analysis completes. That means if an analysis aborts in the middle, this value will not be saved in WfAnalysis even if the analysis is actually for a completed build. This change will save the value at the beginning of the analysis.

Change-Id: I3045ea65fdc16a26faad9569f2f1833a8a63748c
Reviewed-on: https://chromium-review.googlesource.com/c/1257569
Auto-Submit: Chan Li <chanli@chromium.org>
Reviewed-by: Jeffrey Li <lijeffrey@chromium.org>
Commit-Queue: Jeffrey Li <lijeffrey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#18058}