test-results: Marshal everything for test list
Currently, when you request a json response from the /testfile endpoint,
when asking for a list of tests, it only returns the keys. Instead,
just marshal everything about the test files.

This will be used to programmatically determine the name of the layout
test step name for a build, since it can sometimes vary. See the linked
bug for details.

Bug: 882946
Change-Id: I9796fae897228ced30dfa89e6a384d8a88833b77
Reviewed-on: https://chromium-review.googlesource.com/c/1262104
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>
Cr-Commit-Position: refs/heads/master@{#18073}