[tricium spellchecker] Only read the dictionary file once
Previously we were reading the same dictionary.txt file
repeatedly for each file being checked in a CL. The contents
were being saved in a global `dict` variable.

This CL changes it so that we load the dict once, and also
adds a TODO note suggesting to not use the global variable.

Change-Id: I0036e0153275c2b178067b64423e9c647b721813
Reviewed-on: https://chromium-review.googlesource.com/c/1263418
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Cr-Commit-Position: refs/heads/master@{#18074}