[Findit] Flake Analyzer - Include bug id when notifying culprits
Developers shouldn't have to go digging through links to get to the actionable
parts of their work when it comes to fixing flakiness. This change is to fulfill
a feature request to include a link to the bug in the message notification.

Eventually, all bug-related operations will live in a separate layer from flake
analysis, but this API shouldn't need to change much if at all when that time
comes.

Bug: 874999
Change-Id: Ie2df2679681a090f2e2deb3ae72690a95a3ee0a5
Reviewed-on: https://chromium-review.googlesource.com/c/1262918
Commit-Queue: Jeffrey Li <lijeffrey@chromium.org>
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#18078}