[rotang] Handling shifts spanning PDT/PST changes.
The Shift StartTime is in MTV time.
This CL handles changes between PDT/PST; without this change half the
year the shifts will show 23:00 instead of the expected 00:00 as the
start of a shift.

Time is hard.

Additionally the helperFunction adminOrOwner was added to get less repeated code
figuring out if the user owns the configuration to be changed.

Bug: 845549
Change-Id: Iac88d8f25b22008417fa3b098176000b2e7e92fc
Reviewed-on: https://chromium-review.googlesource.com/c/1258727
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Commit-Queue: Ola Karlsson <olakar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#18080}