[third_party_packages_ng] Do not use run.py for dockerbuild.
Dockerbuild works without any of the bootstrap env stuff in infra.git;
bypassing run.py prevents the need for any of that stuff (just the
basic dockerbuild python module).

`run.py` and the additional bootstrap junk is intentionally NOT included
in the recipe bundle, so the recipe currently fails on the bots.

R=hinoka@chromium.org, tandrii@chromium.org, vadimsh@chromium.org

Bug: 877775
Change-Id: Ib0dbdcc86093879e1dd129409584717ef19fcdf6
Reviewed-on: https://chromium-review.googlesource.com/c/1263584
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Cr-Commit-Position: refs/heads/master@{#18083}