[Findit] Fix a gtest normalization bug
Previously, the regular expression used to normalize gtests assume that
the second parameter part can only be a number, however, it's not the
case, for example: ZipFiles/FilesAppBrowserTest.Test/zipCreateFileDrive
is a valid gtest name.

It is impossible to perfectly tell gtests apart from webkit_layout_tests,
for example, a/b.html can be both a valid type parameterized version of
a.html and a webkit_layout_tests, this CL also adds an option to pass
the step name to the normalize function to help identify the type of
the test.

Bug: 891374
Change-Id: Iebc76e36982c70d075b6fd7875a43c5b1af43b28
Reviewed-on: https://chromium-review.googlesource.com/c/1263419
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#18101}