[buildbucket] Handle nonexisting buckets
It is possible that get_accessable_buckets cached a bucket that was later
removed. In such case, get_builders passes id of a non-existing bucket to
config.get_buckets_async, which causes a NotFoundError.

In config.get_buckets_async, return a None for a non-existing bucket.

TBR=vadimsh@chromium.org

Change-Id: Iadc9f9b62e73c00e3ad3a3328f8032b0d0810e90
Reviewed-on: https://chromium-review.googlesource.com/c/1267476
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#18146}