[Findit] Only record once for a completed test heuristic analysis.
The metrics was collected at step level, but when an analysis is aborted, we don't have enough information to decide which step actually causes the analysis to abort. So we should not include step level info to the metrics for now. Later when we change the way to trigger test failure analyses, we can add the step level info then.

Bug: 891412
Change-Id: Ia00dece7576864a839896273103beffda666f52a
Reviewed-on: https://chromium-review.googlesource.com/c/1257573
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Commit-Queue: Chan Li <chanli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#18168}