[Monorail] Fix ts_mon_js_test testbed usage.
A recent CL added a call to the GAE users API, so this test needs to
use testbed to stub that out.  It worked most of the time because some
other test probably leaves the testbed active, but it fails depending
on the order in which tests are run.

R=jeffcarp@chromium.org

Change-Id: I400d321c0703e70bf7103ca2819b3d9d242f0aad
Reviewed-on: https://chromium-review.googlesource.com/c/1270780
Reviewed-by: Joanna Wang <jojwang@chromium.org>
Reviewed-by: Jeff Carpenter <jeffcarp@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Cr-Commit-Position: refs/heads/master@{#18183}