[Monorail] Change direct-link URLs to crbug.com.
This makes it easier for users to get a short crbug link by using
the context menu to copy the link address to the clipboard.

It slows down navigation through a direct-link slightly, but the main
purpose of the direct links is for people to get the URL that they want
onto the clipboard.  If they just want to navigate, then can click
anywhere else on the issue row and not incur a redirect.

R=jojwang@chromium.org

Bug: monorail:3829
Change-Id: I106ab1113b8e82f1699eb2678e86ea437e6b553b
Reviewed-on: https://chromium-review.googlesource.com/c/1272036
Reviewed-by: Joanna Wang <jojwang@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Cr-Commit-Position: refs/heads/master@{#18203}