[Findit] Flake Analyzer - Handler to check recent flakiness
Handler-side implementation to kick off analysis of flakiness at a recent
commit position. Recent commit positions are identified by whatever is
indexed as an IsolatedTarget. If no target exists (due to original analysis
being that of a build not yet on LUCI), buildbot's latest build is used
as a fallback.

Bug: 891915
Change-Id: Ic03946ecd4d1400281897c2cd86c7cf77b086a85
Reviewed-on: https://chromium-review.googlesource.com/c/1260063
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Jeffrey Li <lijeffrey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#18254}