[mr] get #issue_update_form using document.querySelector
document.getElementId gets confused when a form element has an input
child element whose name is 'id' (which this form does contain).

The behavior is to make form input elements reachable in JS by
referring to <form-element>.<input name>.  So
<#input_update_form>.id actually refers to an <input name="id">
element rather than the "id" *attribute* declared on the element in
html.

🤷

getElementById('issue_update_form') still works in Chrome, but FF
returns null. So the init code isn't able to attach to the onsubmit
handler for this form in FF and thus can't do the dirty bit fix, and
thus thinks this is navigating away normally instead of submitting
the form. Hence popup dialog.

I suppose this some how clobbers the 'issue_update_form' entry in
FF's id->element index or something like that

Workaround is to use document.querySelector with a selector that
specifies the id attribute.

Bug: monorail:4320
Change-Id: I2b86587365d7044480764565797b41b3bae4b11a
Reviewed-on: https://chromium-review.googlesource.com/c/1285477
Reviewed-by: Jason Robbins <jrobbins@chromium.org>
Commit-Queue: Sean McCullough <seanmccullough@chromium.org>
Cr-Commit-Position: refs/heads/master@{#18414}