qscheduler: add and implement a NotifyRequest call to scheduler
The NotifyRequest call to the scheduler is used to ensure that scheduler
state is consistent with a request's state at a given time. This is a way
to heal any inconsistencies that may result from dropped pubsub or lost
Assignments between swarming and quotascheduler.

This CL adds the concept of ConfirmedTime to Request and
ConfirmedIdleTime to Worker, which represents the most recent time
at which the state of that record was updated from the authoritative
source (swarming).

BUG=chromium:870537
TEST=Unit test added

Change-Id: Ieb1a177ca1f868bd84450e2150603bde168f1baf
Reviewed-on: https://chromium-review.googlesource.com/c/1262930
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#18421}