[Findit] Update flakes with counts for 'retry_with_patch' flakes as well.
This CL is a quick workaround to count 'retry_with_patch' flakes: it counts flakes of different types separately but it reuses the existing false_rejection_count_last_week as if it doesn't differentiate them from false rejections.

This is to bring back flake data on the dashboard asap.

I'll have a follow-up CL to properly use the counts.

Bug: 896006
Change-Id: I7859608fd09f63c4a898359cc50c67ec447c06ad
Reviewed-on: https://chromium-review.googlesource.com/c/1285471
Commit-Queue: Chan Li <chanli@chromium.org>
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#18439}