qscheduler: treat same-instant NotifyRequest as forward-in-time
Same-instance NotifyRequest calls are unlikely to occur in the wild, but
they do occasionally occur when calling time.Now(). This has been
observed to cause flakiness in the forthcoming doctest in
https://chromium-review.googlesource.com/c/infra/infra/+/1292766

Treat same-instant notifications as though they were forward in time, to
eliminate the flakiness of that test.

BUG=chromium:870537
TEST=New unit tests added, verified to fail prior to fix.

Change-Id: Ia9f5abbf223f6fd344d2e3d561555194e6bf1ae1
Reviewed-on: https://chromium-review.googlesource.com/c/1297522
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Cr-Commit-Position: refs/heads/master@{#18553}