[Findit] Flake Analyzer - Hide check recent flakiness card if main analysis is in progress
The check recent flakiness card will say "Analyzing..." when the main
analysis is in progress. This can be slightly confusing, as it's not clear
what's being analyzed (culprit or recent flakiness). Simply don't show the
check recent flakiness card when the main analysis is in progress.

Bug: 899988
Change-Id: If3e07e85f8a3202a29400a14e0fae4b2006967cc
Reviewed-on: https://chromium-review.googlesource.com/c/1306674
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Jeffrey Li <lijeffrey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#18665}