[Monorail] Fix handling of auxclick events.
The web has evolved to use click and auxclick, so I don't need the old
mouseup and mousedown hack.  Firefox seems to always open a link on
middle-mouse click before my handler even runs, leading to double tabs,
so we are forced to not compute URLs in that case.

R=zhangtiff@chromium.org

Bug: monorail:4451
Change-Id: I47fab5d4b06fa1b26dd0afaf39dba274853bb0cf
Reviewed-on: https://chromium-review.googlesource.com/c/1308781
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Cr-Commit-Position: refs/heads/master@{#18699}