[monorail] Limit conversion and deleting volume.
- Changing the CONVERT_NUM to 20
- Adding a separate VERIFY_NUM = 400 for the Verify() script
- changing UndoConversion() and Conversion() to use visible_results so
changes are limited to 20 issues each time.

The conversion test runs have occasionally resulted in a DeadlineExceededError.
While this still results in correctly converted issues and I can refresh the page
to safely convert the remaining issues, I still want to avoid this and get a safe finish
with the JSON data returned.

I am limiting the number of issues we convert and delete, with each execution, to 20.
Verify() does not have this problem because it's much faster.

Since we can run all scripts as many times as we want, this won't be a problem.
The conversion process will be:
for finch and default => convert 20, verify all, delete (if necessary), repeat until all issues have been converted.

Bug: monorail:4361
Change-Id: I53ccdf82903d5a3c5ce9052ba7b6a606691f8ee7
Reviewed-on: https://chromium-review.googlesource.com/c/1312235
Reviewed-by: Jason Robbins <jrobbins@chromium.org>
Commit-Queue: Joanna Wang <jojwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#18734}