[3pp] SSL hack version 3
Using the power of ctypes, we now load the certs directly from the OS
Keychain via Security.framework. This is a modernized version of what
the `security` tool was doing anyway (SecItemCopyMatching is the
modernized version of SecKeychainSearchCreateFromAttributes, which is
how the `security` tool does it).

Source for `security` tool (see `do_keychain_find_certificate`):
https://opensource.apple.com/source/Security/Security-58286.70.7/SecurityTool/keychain_find.c.auto.html

R=tandrii@chromium.org, vadimsh@chromium.org

Bug: 877775,898348
Change-Id: I43e638f25314209b91eb5aaa94fffcd510f4b7a9
Reviewed-on: https://chromium-review.googlesource.com/c/1315835
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#18773}