qscheduler: initial appengine app for qscheduler-swarming
This CL adds an initial implementation of the qscheduler-swarming
appengine app.

In this implementation, the state of a scheduler is read from and
rewritten to datastore on every client call, inside a transaction. This
is a simple implementation to get us off the ground, but will not scale
to > ~1 QPS due to datastore write contention. Eventually, this approach
will be replaced with some combination of:
 - an in-memory copy of the scheduler state, periodically checkpointed
 into datastore.
 - batching of multiple AssignTasks or NotifyRequest calls for a pool
 into a single transaction.

BUG=chromium:898322
TEST=local `prpc call` tests against dev instance; new basic unit tests
added

Change-Id: I6dadc521ab102bfbb3cc93c97b94601872e63369
Reviewed-on: https://chromium-review.googlesource.com/c/1297520
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#18839}