Fix gatekeeper mailer
It asserts that steps always have a 'results' field. This is not always
true (if the step is still running, for example), and seems to be
unneeded (the step.html template seems to use the value, but given the
current data format, it would throw an exception if it actually used
it, so it's not...?).

Bug: 903421
Change-Id: I81e9c327c8ff0b738cbd61546b13632477ad3a5b
Reviewed-on: https://chromium-review.googlesource.com/c/1327282
Reviewed-by: Ryan Tseng <hinoka@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Cr-Commit-Position: refs/heads/master@{#18876}