[Monorail] Allow UpdateIssue to create comment with just attachments.
There was a condition that prevents the posting of a comment that has
no content and no amendments.  I adjusted it to only prevent posting
if there are also no attachments.  So, the case where the user wants
to just post an attachment goes through.

R=zhangtiff@chromium.org

Bug: monorail:4429
Change-Id: I717bcf6b477084e11c8d4f48cf567949b9d7d0d1
Reviewed-on: https://chromium-review.googlesource.com/c/1324171
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Cr-Commit-Position: refs/heads/master@{#18892}