[kitchen] Stream steps and properties
Add AnnotationUpdated callback that converts the updated annotation proto
into an UpdateBuild RPC and streams updated steps and output properties
to buildbucket.

Add -buildbucket-build-id, since UpdateBuild RPC requires it.

Change runBuildUpdater to exit when context is canceled, not when
requests channel is closed. We must exit when cancel is cancelled
(since we cannot make RPCs with a cancelled context), and it is an
unnecessary complexity to have two exit points.

Bug: 843714
Change-Id: I0640ab91e5244af181da1b60eb078a90a11ccdf9
Reviewed-on: https://chromium-review.googlesource.com/c/1329952
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Cr-Commit-Position: refs/heads/master@{#18925}