[Monorail] Make keyword "me" handle linked accounts.
Queries for terms like [owner:me], [cc:me], and [is:starred] are
translated into lower-level query terms that specify the user ID of
the user involved: Usually that is the signed in user, but it can also
be the viewed_user when viewing another user's dashboard (a feature that
is still on on hiatus).  With this change, the single user ID is replaced
with a list of IDs of all the accounts linked to the specified account.
E.g., searching for [owner:me] as jrobbins@chromium.org would show
issues owned by either jrobbins@chormium.org or jrobbins@google.com.

R=jeffcarp@chromium.org

Bug: monorail:94
Change-Id: I314fc87b3a2d84aaad9f150a18bbcd8a66080fb9
Reviewed-on: https://chromium-review.googlesource.com/c/1332797
Reviewed-by: Jeff Carpenter <jeffcarp@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Cr-Commit-Position: refs/heads/master@{#18963}