build_from_tarball: Stop bootstrapping GN with --with-sysroot on recent M72 builds
Following https://gn-review.googlesource.com/c/gn/+/3200,
https://chromium-review.googlesource.com/c/1333391 removed the
sysroot-related options from GN's bootstrap.py in the Chromium tree.

build_from_tarball.py was not updated accordingly, and was failing with
errors such as
  <https://ci.chromium.org/p/infra/builders/luci.infra.cron/Build%20From%20Tarball/b8929434915133491680>

We cannot just stop passing --with-sysroot to bootstrap.py altogether
though, as we still have M70 and M71 tarballs created with a GN without
sysroot options but lacking the corresponding change to bootstrap.py (in
which case not passing --with-sysroot would cause bootstrap.py to call
gen.py with the --no-sysroot option and fail). Yes, this is confusing.

Bug: 907025
Change-Id: I91d03de1c5b0a53c9075bf7c74b5fc3bca128bbb
Reviewed-on: https://chromium-review.googlesource.com/c/1343319
Commit-Queue: Michael Moss <mmoss@chromium.org>
Reviewed-by: Michael Moss <mmoss@chromium.org>
Cr-Commit-Position: refs/heads/master@{#19092}