[code coverage] Serve rebased flat data for per cl requests
The correct way to do this is to change everything to the compressed
format, and that would require non-trivial refactoring and changes at
the recipe side, and that's a much bigger efforts than this CL intends
to cover.

So, this CL first implements this workaround at the service side, and
then will switch to compressed format as part of the efforts to
eliminate flat form, tracked in: crbug.com/904469.

Change-Id: Id9f0692bd8103a770fe867a7037b4eacb3c9b512
Reviewed-on: https://chromium-review.googlesource.com/c/1343432
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#19110}