[dockerbuild] Fix lacking autogen.sh invocation in libffi build
The libffi source changed (because of my own stupidity), and I didn't
update the dockcross scripts to account for that.

In particular, the version of the libffi source previously uploaded
was a derived 'dist' version which didn't work with OS X. I re-uploaded
the non-derived source tarball, but it requires an 'automake' step which
the baked distribution doesn't (incedentally, this is the reason why the
baked version DOESN'T work on OS X).

R=tandrii@chromium.org, vadimsh@chromium.org

Bug: 877775
Change-Id: I074b7a167e2562ae936af41e33a8c1242b909d6a
Reviewed-on: https://chromium-review.googlesource.com/c/1347303
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#19134}