[Monorail] Fix sorting by owner.
Sorting by owner worked in small projects, but it did not work in
/p/chromium in prod because of DB query timeouts.  The cause seems to
be that MySQL will not use our index on a LEFT JOIN that has an
OR operator in the join condition.  Using 2 LEFT JOINS works.

Time for the SQL query for an issue list sorted by owner is
reduced from over 70 seconds to under 1 second.  Testing on
staging in /p/chromium went from timeouts to rendering the
issue list in about 9 seconds, and /p/monorail went from 5 seconds
to 1 second.

R=jeffcarp@chromium.org

Bug: monorail:1955
Change-Id: I0205df8644beb6b579aa83aeea80c22aafc5b068
Reviewed-on: https://chromium-review.googlesource.com/c/1337538
Reviewed-by: Jeff Carpenter <jeffcarp@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Cr-Commit-Position: refs/heads/master@{#19165}