[Code Coverage] Handle the sharded coverage data.
Now the full-repository coverage provides sharded output for files, this CL is
to process the data in the new structure to avoid OOM on Google App Engine.
Originally, processing 150MB raw metadata in json format will consume 1.8GB
memory and that hits the 1GB memory cap on GAE.

With the sharded output being downloaded and processed one by one, the memory
usage is under ~500MB, and this should be safe on GAE.
Regarding speed, now it takes ~4 minutes to process all the data for a single
full-repository report. This is fast enough, and under the 10-minute limit for
a task in GAE taskqueue.

Change-Id: I8cae2f8c05c288d7c180cede0d08f3cdb7187838
Reviewed-on: https://chromium-review.googlesource.com/c/1354046
Commit-Queue: Shuotao Gao <stgao@chromium.org>
Reviewed-by: Roberto Carrillo <robertocn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#19210}