[tricium] Clarify gerrit/poll.go
When trying to test eslint, I found that my changes to the playground
repo are no longer analyzed after I added a whitelist to that repo,
which means that I either specified the whitelist wrong or there's a
bug in handling of whitelists.

At first I didn't realize that this was the issue so I tried to debug
by reading through poll.go again, and made several changes to try to
clarify what was happening during the poll.

In this change:
 - Update logging to be less verbose and more useful.
 - Extract helper function isOpen, and note that
   we only ever see NEW, and not DRAFT. This is a minor
   change that should not affect actual behavior --
   but it could be separated from this CL.
 - Update comments.

Bug: 908647
Change-Id: Id7572d909900b46c4da96c07bd9bd0b1acb8d4ba
Reviewed-on: https://chromium-review.googlesource.com/c/1351832
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#19211}