[tricium eslint] Support reading larger eslint output
Previously we used bufio.Scanner to read the command output,
since this is how the pylint analyzer was written, because
in that case we read one line at a time, and lines are not
very large.

However in this case we want to read the whole command output
at once, not line by line, since eslint has JSON output, and
the JSON output is one big JSON blob.

Bug: 910746
Change-Id: Ibd86923b107c74c647f88c7e3ae44c60efa78921
Reviewed-on: https://chromium-review.googlesource.com/c/1356853
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#19298}