qscheduler: rename some proto messages and enums for clarity
In particular, rename reconciler.TaskUpdate to reconciler.TaskInstant,
to emphasize that it represents the state of a particular task at a
particular time. While at it, rename the enum values that express a
state to WAITING, RUNNING, and ABSENT which are better descriptions of
the states with respect to reconciler/scheduler.

No behavioral change (other than to remove a case that was already not
being used).

Change-Id: I8b07824dc86f27b424fa354bca1f41089a6cd181
Reviewed-on: https://chromium-review.googlesource.com/c/1357939
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#19312}