[tricium] Poll for each project in a separate request
Currently, all projects are polled in parallel in one request,
which means that having one large repo (like chromium/src) may
cause the request to fail due to out of memory etc.

In general, in the future we want to adopt a "multi-tenant" approach
wherein operations for separate LUCI projects are always kept
separate.

So, to bring the Gerrit polling in line with this approach,
to scale up, and to avoid OOM due to one heavy project, this CL
would split polling by project.

Specifically, the poll handler now just puts "poll-project"
requests in a queue, which are then picked up, and each of those
requests polls for each repo in the project.

Bug: 908636
Change-Id: If85c595bc1910d624897511573835b17ceeb8206
Reviewed-on: https://chromium-review.googlesource.com/c/1372741
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#19519}