[tricium] Simplify analyzer logging.
This simplifies the logging in all of the Go analyzers,
mainly by not printing the input FILES and output RESULTS,
which are already viewable on isolateserver for jobs in
production or on the filesystem locally for tests.

The purpose is to make it easier to look at the logs
and debug in the future.

Change-Id: I5cbba95a1a6d0eb8dc0428f9c2f4e3c87d0ca71f
Reviewed-on: https://chromium-review.googlesource.com/c/1376972
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#19573}