[tricium] Limit polling to one request with a fixed number of changes
Specifically, this CL changes the limit to 60 per minute, as an
initial number that could be adjusted later.

For smaller projects there tends to be only a few changes per minute
at most, so this will always be more than enough.  For Chromium,
most of the time it appears that there is also usually only a few
published changes per minute, so this will also be sufficient for
chromium/src, even considering that we include WIP CLs.
(https://chromium-review.googlesource.com/q/project:chromium%252Fsrc+status:open)

In the rare event of a burst of >60 changes in a minute, it may
be good to limit requests anyway, to avoid OOM (which is what is
happening currently). We also hope to avoid too many requests to
Gerrit anyway, and a limit of one request per minute per project
seems reasonable.

The expected effect of this CL is that the polling of chromium/src
should succeed, and the GerritProject entity will be updated,
so then polling will start working on current CLs.

Bug: 908636
Change-Id: Ibfc40483473d4606a714fe80472ea65a32ee8fc9
Reviewed-on: https://chromium-review.googlesource.com/c/1378746
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Cr-Commit-Position: refs/heads/master@{#19614}