qscheduler: decouple qslib API from State proto representation
No behavioral change, though some of the tests had to be modified
because they relied on proto representation.

The API still depends on Config proto representation, though the
dependence is not as intertwined. This will be fixed in a future CL.

This (in the end, much bigger than I wanted) CL extracts a lot of the
proto depenence out of the qslib API, deleting several modules and
moving a bunch of code around in the process.

BUG=chromium:907292
TEST=tests pass

Change-Id: I8fe607ce9b43ad690c296e91c8078bf4617aa6bd
Reviewed-on: https://chromium-review.googlesource.com/c/1378796
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>
Cr-Commit-Position: refs/heads/master@{#19643}