[Findit] Flake detection -- detect hidden flakes due to test retries in test runner.
A hidden flake is a test that, within a single test step:
* Fail in 2+ runs (including the first failed run and retries by the test runner)
* Pass at least 1 run

This implies that a hidden flake is retried at least twice. This is because a
test could flake due to resource starvation when the test runner is in parallel
test execution mode.

Change-Id: Ibee620f5fe3d9928658a0d76f534d965d406e49c
Reviewed-on: https://chromium-review.googlesource.com/c/1382605
Commit-Queue: Shuotao Gao <stgao@chromium.org>
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#19647}