[tricium] In Progress RPC, use status OK for "no run".
The rationale here is that "no run" is the normal case, and it doesn't
mean that client did anything wrong, so it shouldn't be a 4xx error.
The client can easily see that there is no run because the response
is an empty response (with no run ID set) in this case.

This would also make the monitoring change for https://crbug/876124
simpler, and it would make it easier to find real client errors.

Also in this CL:
 - Change the logging in this case to be the URL of the change,
   which is more useful for debugging.
 - Remove a useless debug log line.

Bug: 917500
Change-Id: I4e66b1cd6fabe0434410229b67ad5df3966c1419
Reviewed-on: https://chromium-review.googlesource.com/c/1389217
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#19716}