[Monorail] ClientLogger: return early instead of throwing for stale input
Previously ClientLogger would throw an error if it encountered events with
no label or missing starting entries. This is our fault since it's possible
that the sessionStorage has been cleared between the time we start and pause
or resume timers. ClientLogger throwing unhandled JS errors is a problem
because it can cause other JS to fail. See bug for more details. This CL updates
ClientLogger to simply return early instead of throwing an error.

Bug: monorail:4698
Change-Id: I50f0adbb30a8d794a4f0bcadcc02034031c4f096
Reviewed-on: https://chromium-review.googlesource.com/c/1382610
Commit-Queue: Jeff Carpenter <jeffcarp@chromium.org>
Reviewed-by: Edward Lesmes <ehmaldonado@chromium.org>
Cr-Commit-Position: refs/heads/master@{#19750}