[tricium] Temp debugging change: list membership groups
This should have no effect on behavior.

From previous experiments, I saw that checking the groups
 - project-fuchsia-committers,
 - project-chromium-tryjob-access, and
 - all
all seem to work as expected, whereas the groups
 - project-infra-committers, and
 - project-infra-tryjob-access
both seem to never contain qyearsley@chromium.org.

To continue debugging this, as a next step, I plan
to add more whitelisted groups to the playground config
and see which ones seem to work as expected. Maybe
there will be some kind of pattern.

Meanwhile, this change adds extra logging to regular
requests, so it should be reverted when the mystery
is solved.

TBR=maruel
TBR_REASON=temporary change that should not affect behavior

Bug: 908647
Change-Id: I513e9c6f724d17cf2acfed0e672b0319de60423a
Reviewed-on: https://chromium-review.googlesource.com/c/1391369
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Cr-Commit-Position: refs/heads/master@{#19766}