[Findit] Fix the breakage caused by disabling "retry with patch".
Recently, a support was added to disable "retry with patch", and GPU tests opted
out to not run "retry with patch" (https://crbug.com/906166). This caused a
2-week breakage in flake detection for GPU tests.

This CL is to make the queries support test steps with "retry with patch"
enabled or disabled.

Bug: 916762
Change-Id: I5015d166aa04a45c21a4f9b6c1424dae8336010b
Reviewed-on: https://chromium-review.googlesource.com/c/1388674
Commit-Queue: Shuotao Gao <stgao@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#19793}