[kitchen] Regen recipe_engine protos to remove deprecated UseResultProto.
This is the last vestige of the 'use_result_proto' flag. The recipe
engine now does this by default and ignores the flag, so remove it
from kitchen so that we can finally remove it from the recipe engine
code.

R=martiniss@chromium.org, nodir@chromium.org

Bug: 646142
Change-Id: Ia4e5274819cc149a88c4732ddc8154652e16ec9e
Reviewed-on: https://chromium-review.googlesource.com/c/1402321
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Cr-Commit-Position: refs/heads/master@{#19853}