qscheduler: delete LabelSet struct
LabelSet was intended as a optimized version of a string set that works
particularly well for length 1 sets. However, the concept of label
matching within qscheduler is going to grow to include all task and bot
labels, not just provisionable labels, so the length 1 optimizaion is
much less valuable.

BUG=chromium:920352

Change-Id: I333a1a2450424b7cc4aca0078cf664ad59462ee9
Reviewed-on: https://chromium-review.googlesource.com/c/1405617
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Cr-Commit-Position: refs/heads/master@{#19938}