skylab: create-test: add -keyval argument
This adds a -keyval argument to create-test, allowing users to specify
autotest test keyvals. It uses -keyval KEY:VALUE format which is a lot
more convenient than forcing users to input a json string. (note: this
will not support keyvals where there is a ":" character in the key, but
it's not clear to me that we have any of these).

BUG=919258

Change-Id: I7081aa38e54feb0f2ada348f673f8479ded790f1
Reviewed-on: https://chromium-review.googlesource.com/c/1415384
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#20055}