qscheduler: support heterogenous matching
Prior to this, qscheduler assumed that (other than provisionable labels)
all the workers in its pool were interchangable with respect to any of
the tasks in the pool.

This requirement is too strict for use by test planning engine.

This CL eliminates that requirement, and significantly restructures the
implementation to accomplish this. In particular:

 - Scheduling passes are now Worker-oriented, rather than
 Request-oriented.
 - A match heuristic is introduced that attempts to find the best
 matching request for a given worker, by selecting the most restrictive
 request which can run on that worker.
 - A schedulerRun struct is introduced to hold useful intermediate
 state used during a scheduler run.

BUG=920352

Change-Id: Ib4d457f7040c2003d4ac682306d8636b4a35bf96
Reviewed-on: https://chromium-review.googlesource.com/c/1407880
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#20120}