qscheduler-swarming: factor state modification logic out of rpc handlers
This CL factors out the logic for handling an individual AssignTasks
or NotifyTasks call from the appending handlers and separates it from
the datastore operations to load/persist state.

This lays the groundwork for an alternative implementation of the server
that batches multiple request calls into a single datastore transaction.

BUG=923515
TEST=existing tests pass

Change-Id: I177eb057cc5e387de730b855bfd402d468fa6e82
Reviewed-on: https://chromium-review.googlesource.com/c/1422780
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#20198}