skylab_swarming_worker: Reorder harness steps
Reorder steps so the cleanup is the exact reverse of the setup.

Fix a minor bug where we close the results dir before closing the
hostinfo, which technically results in a race where gs_offloader could
offload the results before we read back the hostinfo.  (In practice
gs_offloader is nowhere fast enough to do this.)

R=akeshet@chromium.org

Change-Id: Ice8a0fbd3dd24ab1a323ee166f581fe3cf04ceb9
Reviewed-on: https://chromium-review.googlesource.com/c/1437946
Commit-Queue: Allen Li <ayatane@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Cr-Commit-Position: refs/heads/master@{#20323}